+++ agent/temboardagent/plugins/dashboard/metrics.py
@@ -22,6 +22,7 @@ def get_metrics(app):
                 max_connections=dm.get_max_connections(),
                 databases=dm.get_stat_db(),
                 pg_uptime=dm.get_pg_uptime(),
+                pg_start_time=dm.get_pg_start_time(),
                 pg_version=pginfo.version()['full'],
                 pg_data=pginfo.setting('data_directory'),
                 pg_port=pginfo.setting('port'),
@@ -88,6 +89,7 @@ def get_info(conn, config):
         hostname=sysinfo.hostname(config.temboard.hostname),
         os_version=' '.join([sysinfo.os, sysinfo.os_release]),
         pg_uptime=dm.get_pg_uptime(),
+        pg_start_time=dm.get_pg_start_time(),
         pg_version=pginfo.version()['full'],
         pg_data=pginfo.setting('data_directory'),
         pg_port=pginfo.setting('port'),
@@ -220,6 +222,9 @@ class DashboardMetrics:
                 'total_commit': row['total_commit'],
                 'total_rollback': row['total_rollback'],
                 'timestamp': time.time()}
+
+    def get_pg_start_time(self):
+        return self.conn.queryscalar("SELECT pg_postmaster_start_time();")
 
     def get_pg_uptime(self,):
         return self.conn.queryscalar("""\